fix(kserve-module): move drift test to lifecycle and replace sleep with polling (#1621)
Move TestDriftCorrection from test_status.py to test_lifecycle.py where
it belongs — drift correction is reconciler behavior, not status reporting.

Replace time.sleep(15) with wait_for() polling helper to avoid flaky
failures on slower clusters (e.g. OCP) where reconcile takes longer
than the fixed sleep window.
