+++ kserve-module/tests/e2e/test_status.py
@@ -1,4 +1,4 @@
-"""E2E tests for status conditions and drift correction.
+"""E2E tests for status conditions.
 
 Deployment unavailability, error isolation, and dependency handling are covered
 by integration tests (reconciler_int_test.go, dependency_int_test.go) because
@@ -6,22 +6,12 @@ SSA re-applies desired state before the readiness check runs, making it
 impossible to simulate in E2E.
 """
 
-import time
-
 import pytest
 
 from conftest import (
-    run,
     get_cr,
-    trigger_reconcile,
-    NAMESPACE,
+    get_conditions,
 )
-
-
-def _get_conditions(kubectl):
-    """Fetch conditions as a dict keyed by condition type."""
-    cr = get_cr(kubectl)
-    return {c["type"]: c for c in cr.get("status", {}).get("conditions", [])}
 
 
 @pytest.mark.sanity
@@ -30,7 +20,7 @@ class TestStatusConditions:
 
     def test_happy_path_all_conditions(self, kubectl, cluster_info, apply_kserve_cr):
         """All conditions report correctly after successful reconcile."""
-        conditions = _get_conditions(kubectl)
+        conditions = get_conditions(kubectl)
 
         assert conditions["Ready"]["status"] == "True"
         assert conditions["ProvisioningSucceeded"]["status"] == "True"
@@ -50,34 +40,3 @@ class TestStatusConditions:
         assert cr["status"]["observedGeneration"] == cr["metadata"]["generation"]
 
 
-@pytest.mark.sanity
-class TestDriftCorrection:
-    """SSA drift correction — manual edits reverted on reconcile."""
-
-    def test_configmap_edit_reverted(self, kubectl, apply_kserve_cr):
-        """Manual ConfigMap edit is reverted by SSA on next reconcile."""
-        cm_name = "inferenceservice-config"
-
-        run([
-            kubectl, "patch", "configmap", cm_name, "-n", NAMESPACE,
-            "--type", "merge",
-            "-p", '{"data":{"ingress":"{\\"ingressClassName\\":\\"TAMPERED\\"}"}}',
-        ])
-
-        result = run([
-            kubectl, "get", "configmap", cm_name, "-n", NAMESPACE,
-            "-o", "jsonpath={.data.ingress}",
-        ])
-        assert "TAMPERED" in result.stdout
-
-        trigger_reconcile(kubectl, trigger_id="drift-001")
-        time.sleep(15)
-
-        result = run([
-            kubectl, "get", "configmap", cm_name, "-n", NAMESPACE,
-            "-o", "jsonpath={.data.ingress}",
-        ])
-        assert "TAMPERED" not in result.stdout
-
-        conditions = _get_conditions(kubectl)
-        assert conditions["ProvisioningSucceeded"]["status"] == "True"
